update exception handling
