+++ tests/middleware/test_middleware_handling.py
@@ -111,8 +111,8 @@ def test_setting_cors_middleware() -> None:
             cur = cast(ASGIApp, cur.app)  # type: ignore
         else:
             unpacked_middleware.append(cur)
-        assert len(unpacked_middleware) == 2
-        cors_middleware = unpacked_middleware[0]
+        assert len(unpacked_middleware) == 4
+        cors_middleware = unpacked_middleware[1]
         assert isinstance(cors_middleware, CORSMiddleware)
         assert cors_middleware.allow_headers == ["*", "accept", "accept-language", "content-language", "content-type"]
         assert cors_middleware.allow_methods == ("DELETE", "GET", "HEAD", "OPTIONS", "PATCH", "POST", "PUT")
@@ -132,8 +132,8 @@ def test_trusted_hosts_middleware() -> None:
         cur = cast(ASGIApp, cur.app)  # type: ignore
     else:
         unpacked_middleware.append(cur)
-    assert len(unpacked_middleware) == 2
-    trusted_hosts_middleware = unpacked_middleware[0]
+    assert len(unpacked_middleware) == 4
+    trusted_hosts_middleware = unpacked_middleware[1]
     assert isinstance(trusted_hosts_middleware, TrustedHostMiddleware)
     assert trusted_hosts_middleware.allowed_hosts == ["*"]
